Conflicts:
	pyqtgraph/imageview/ImageView.py - nframes v. Key.*

+++ examples/test_examples.py
@@ -1,16 +1,13 @@
-# -*- coding: utf-8 -*-
-from __future__ import print_function, division, absolute_import
 from collections import namedtuple
 from pyqtgraph import Qt
-
 import errno
+import time
 import importlib
 import itertools
 import pytest
 import os, sys
 import platform
 import subprocess
-import time
 from argparse import Namespace
 if __name__ == "__main__" and (__package__ is None or __package__==''):
     parent_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
@@ -34,7 +31,7 @@ def buildFileList(examples, files=None):
 
 
 path = os.path.abspath(os.path.dirname(__file__))
-files = [("Example App", "test_ExampleApp.py")]
+files = [("Example App", "RunExampleApp.py")]
 for ex in [utils.examples, utils.others]:
     files = buildFileList(ex, files)
 files = sorted(set(files))
@@ -56,12 +53,15 @@ installedFrontends = sorted([
     frontend for frontend, isPresent in frontends.items() if isPresent
 ])
 
+darwin_opengl_broken = (platform.system() == "Darwin" and
+            tuple(map(int, platform.mac_ver()[0].split("."))) >= (10, 16) and
+            sys.version_info < (3, 9, 1))
+
+darwin_opengl_reason = ("pyopenGL cannot find openGL library on big sur: "
+                        "https://github.com/python/cpython/pull/21241")
+
 exceptionCondition = namedtuple("exceptionCondition", ["condition", "reason"])
 conditionalExamples = {
-    "test_ExampleApp.py": exceptionCondition(
-        not(platform.system() == "Linux" and frontends[Qt.PYSIDE2]),
-        reason="Unexplained, intermittent segfault and subsequent timeout on CI"
-    ),
     "hdf5.py": exceptionCondition(
         False,
         reason="Example requires user interaction"
@@ -71,104 +71,48 @@ conditionalExamples = {
         reason="Test is being problematic on CI machines"
     ),
     'GLVolumeItem.py': exceptionCondition(
-        not(platform.system() == "Darwin" and
-            tuple(map(int, platform.mac_ver()[0].split("."))) >= (10, 16) and 
-            (sys.version_info <= (3, 8, 7) or
-            (sys.version_info >= (3, 9) and sys.version_info < (3, 9, 1)))),
-        reason=(
-            "pyopenGL cannot find openGL libray on big sur: "
-            "https://github.com/python/cpython/pull/21241"
-        )
+        not darwin_opengl_broken,
+        reason=darwin_opengl_reason
     ),
     'GLIsosurface.py': exceptionCondition(
-        not(platform.system() == "Darwin" and
-            tuple(map(int, platform.mac_ver()[0].split("."))) >= (10, 16) and 
-            (sys.version_info <= (3, 8, 7) or
-            (sys.version_info >= (3, 9) and sys.version_info < (3, 9, 1)))),
-        reason=(
-            "pyopenGL cannot find openGL libray on big sur: "
-            "https://github.com/python/cpython/pull/21241"
-        )
+        not darwin_opengl_broken,
+        reason=darwin_opengl_reason
     ),
     'GLSurfacePlot.py': exceptionCondition(
-        not(platform.system() == "Darwin" and
-            tuple(map(int, platform.mac_ver()[0].split("."))) >= (10, 16) and 
-            (sys.version_info <= (3, 8, 7) or
-            (sys.version_info >= (3, 9) and sys.version_info < (3, 9, 1)))),
-        reason=(
-            "pyopenGL cannot find openGL libray on big sur: "
-            "https://github.com/python/cpython/pull/21241"
-        )
+        not darwin_opengl_broken,
+        reason=darwin_opengl_reason
     ),
     'GLScatterPlotItem.py': exceptionCondition(
-        not(platform.system() == "Darwin" and
-            tuple(map(int, platform.mac_ver()[0].split("."))) >= (10, 16) and 
-            (sys.version_info <= (3, 8, 7) or
-            (sys.version_info >= (3, 9) and sys.version_info < (3, 9, 1)))),
-        reason=(
-            "pyopenGL cannot find openGL libray on big sur: "
-            "https://github.com/python/cpython/pull/21241"
-        )
+        not darwin_opengl_broken,
+        reason=darwin_opengl_reason
     ),
     'GLshaders.py': exceptionCondition(
-        not(platform.system() == "Darwin" and
-            tuple(map(int, platform.mac_ver()[0].split("."))) >= (10, 16) and 
-            (sys.version_info <= (3, 8, 7) or
-            (sys.version_info >= (3, 9) and sys.version_info < (3, 9, 1)))),
-        reason=(
-            "pyopenGL cannot find openGL libray on big sur: "
-            "https://github.com/python/cpython/pull/21241"
-        )
+        not darwin_opengl_broken,
+        reason=darwin_opengl_reason
+    ),
+    'GLTextItem.py': exceptionCondition(
+        not darwin_opengl_broken,
+        reason=darwin_opengl_reason
     ),
     'GLLinePlotItem.py': exceptionCondition(
-        not(platform.system() == "Darwin" and
-            tuple(map(int, platform.mac_ver()[0].split("."))) >= (10, 16) and 
-            (sys.version_info <= (3, 8, 7) or
-            (sys.version_info >= (3, 9) and sys.version_info < (3, 9, 1)))),
-        reason=(
-            "pyopenGL cannot find openGL libray on big sur: "
-            "https://github.com/python/cpython/pull/21241"
-        )
+        not darwin_opengl_broken,
+        reason=darwin_opengl_reason
     ),
     'GLMeshItem.py': exceptionCondition(
-        not(platform.system() == "Darwin" and
-            tuple(map(int, platform.mac_ver()[0].split("."))) >= (10, 16) and 
-            (sys.version_info <= (3, 8, 7) or
-            (sys.version_info >= (3, 9) and sys.version_info < (3, 9, 1)))),
-        reason=(
-            "pyopenGL cannot find openGL libray on big sur: "
-            "https://github.com/python/cpython/pull/21241"
-        )
+        not darwin_opengl_broken,
+        reason=darwin_opengl_reason
     ),
     'GLImageItem.py': exceptionCondition(
-        not(platform.system() == "Darwin" and
-            tuple(map(int, platform.mac_ver()[0].split("."))) >= (10, 16) and 
-            (sys.version_info <= (3, 8, 7) or
-            (sys.version_info >= (3, 9) and sys.version_info < (3, 9, 1)))),
-        reason=(
-            "pyopenGL cannot find openGL libray on big sur: "
-            "https://github.com/python/cpython/pull/21241"
-        )
+        not darwin_opengl_broken,
+        reason=darwin_opengl_reason
     ),
     'GLBarGraphItem.py': exceptionCondition(
-        not(platform.system() == "Darwin" and
-            tuple(map(int, platform.mac_ver()[0].split("."))) >= (10, 16) and
-            (sys.version_info <= (3, 8, 7) or
-            (sys.version_info >= (3, 9) and sys.version_info < (3, 9, 1)))),
-        reason=(
-            "pyopenGL cannot find openGL libray on big sur: "
-            "https://github.com/python/cpython/pull/21241"
-        )
+        not darwin_opengl_broken,
+        reason=darwin_opengl_reason
     ),
     'GLViewWidget.py': exceptionCondition(
-        not(platform.system() == "Darwin" and
-            tuple(map(int, platform.mac_ver()[0].split("."))) >= (10, 16) and
-            (sys.version_info <= (3, 8, 7) or
-            (sys.version_info >= (3, 9) and sys.version_info < (3, 9, 1)))),
-        reason=(
-            "pyopenGL cannot find openGL libray on big sur: "
-            "https://github.com/python/cpython/pull/21241"
-        )
+        not darwin_opengl_broken,
+        reason=darwin_opengl_reason
     )
 }
 
@@ -200,52 +144,46 @@ conditionalExamples = {
     ]
 )
 def testExamples(frontend, f):
-    # runExampleFile(f[0], f[1], sys.executable, frontend)
-
     name, file = f
     global path
     fn = os.path.join(path, file)
     os.chdir(path)
-    sys.stdout.write("{} ".format(name))
+    sys.stdout.write(f"{name}")
     sys.stdout.flush()
     import1 = "import %s" % frontend if frontend != '' else ''
     import2 = os.path.splitext(os.path.split(fn)[1])[0]
     code = """
 try:
-    %s
+    {0}
     import initExample
     import pyqtgraph as pg
-    import %s
+    import {1}
     import sys
     print("test complete")
     sys.stdout.flush()
-    import time
-    while True:  ## run a little event loop
-        pg.QtGui.QApplication.processEvents()
-        time.sleep(0.01)
+    pg.Qt.QtCore.QTimer.singleShot(1000, pg.Qt.QtWidgets.QApplication.quit)
+    pg.exec()
+    names = [x for x in dir({1}) if not x.startswith('_')]
+    for name in names:
+        delattr({1}, name)
 except:
     print("test failed")
     raise
 
-""" % (import1, import2)
-    if sys.platform.startswith('win'):
-        process = subprocess.Popen([sys.executable],
-                                    stdin=subprocess.PIPE,
-                                    stderr=subprocess.PIPE,
-                                    stdout=subprocess.PIPE)
-    else:
-        process = subprocess.Popen(['exec %s -i' % (sys.executable)],
-                                   shell=True,
-                                   stdin=subprocess.PIPE,
-                                   stderr=subprocess.PIPE,
-                                   stdout=subprocess.PIPE)
-    process.stdin.write(code.encode('UTF-8'))
+""".format(import1, import2)
+    process = subprocess.Popen([sys.executable],
+                                stdin=subprocess.PIPE,
+                                stderr=subprocess.PIPE,
+                                stdout=subprocess.PIPE,
+                                text=True)
+    process.stdin.write(code)
     process.stdin.close()
+
     output = ''
     fail = False
     while True:
         try:
-            c = process.stdout.read(1).decode()
+            c = process.stdout.read(1)
         except IOError as err:
             if err.errno == errno.EINTR:
                 # Interrupted system call; just try again.
@@ -253,24 +191,37 @@ except:
             else:
                 raise
         output += c
-
         if output.endswith('test complete'):
             break
         if output.endswith('test failed'):
             fail = True
             break
-    time.sleep(1)
-    process.kill()
-    #res = process.communicate()
-    res = (process.stdout.read(), process.stderr.read())
+    start = time.time()
+    killed = False
+    while process.poll() is None:
+        time.sleep(0.1)
+        if time.time() - start > 2.0 and not killed:
+            process.kill()
+            killed = True
+
+    stdout, stderr = (process.stdout.read(), process.stderr.read())
+    process.stdout.close()
+    process.stderr.close()
+
     if (fail or
-        'exception' in res[1].decode().lower() or
-        'error' in res[1].decode().lower()):
-        print(res[0].decode())
-        print(res[1].decode())
-        pytest.fail("{}\n{}\nFailed {} Example Test Located in {} "
-            .format(res[0].decode(), res[1].decode(), name, file),
-            pytrace=False)
+        'Exception:' in stderr or
+        'Error:' in stderr):
+        if (not fail 
+            and name == "RemoteGraphicsView" 
+            and "pyqtgraph.multiprocess.remoteproxy.ClosedError" in stderr):
+            # This test can intermittently fail when the subprocess is killed
+            return None
+        print(stdout)
+        print(stderr)
+        pytest.fail(
+            f"{stdout}\n{stderr}\nFailed {name} Example Test Located in {file}",
+            pytrace=False
+        )
 
 if __name__ == "__main__":
     pytest.cmdline.main()
